beta4 Radio ready for test

+++ src/common.h
@@ -80,7 +80,7 @@ extern bool     do_log_rf24;
 extern String   rf24gw_hub_server;
 extern uint16_t rf24gw_hub_port;
 extern uint16_t rf24gw_gw_port;
-extern uint8_t  rf24gw_gw_no;
+extern uint16_t rf24gw_gw_no;
 
 void rf24gw_setup();
 void rf24gw_loop();
